+++ scripts/update-dependencies.rb
@@ -1,8 +1,6 @@
 #!/usr/bin/env ruby
-require 'bumpsnag'
-
-target_submodule = ENV['TARGET_SUBMODULE']
-target_version = ENV['TARGET_VERSION']
+target_submodule = ENV['TARGET_SUBMODULE'].dup
+target_version = ENV['TARGET_VERSION'].dup
 
 if target_submodule.nil? || target_version.nil?
   raise 'Submodule or version targets not provided, exiting'
@@ -10,6 +8,8 @@ if target_submodule.nil? || target_version.nil?
 end
 
 pp "Attempting upgrade of #{target_submodule} to #{target_version}"
+
+target_version.delete_prefix!('v') if target_version.start_with?('v')
 
 if target_submodule.eql?('bugsnag-android')
   `packages/react-native/update-android.sh --version #{target_version}`