+++ test/aws-lambda/features/steps/aws-lambda-steps.rb
@@ -1,7 +1,13 @@
 Given('I setup the environment') do
+  if ENV['NETWORK_NAME']
+    host_name = 'host.docker.internal'
+  else
+    # Use the default for docker bridge host
+    host_name = '172.17.0.1'
+  end
   steps %Q{
     Given I store the api key in the environment variable "BUGSNAG_API_KEY"
-    And I set environment variable "BUGSNAG_NOTIFY_ENDPOINT" to "http://host.docker.internal:9339/notify"
-    And I set environment variable "BUGSNAG_SESSIONS_ENDPOINT" to "http://host.docker.internal:9339/sessions"
+    And I set environment variable "BUGSNAG_NOTIFY_ENDPOINT" to "http://#{host_name}:9339/notify"
+    And I set environment variable "BUGSNAG_SESSIONS_ENDPOINT" to "http://#{host_name}:9339/sessions"
   }
 end